+++ Thunderstore/CHANGELOG.md
@@ -1,2 +1,6 @@
+## v1.1.0
+- Added the possibility to also disable items in the lobby
+- optimized & cleaned up code
+
 ## v1.0.0
 - Initial Release